add navigation to combobox on editor
